requirements.txt
python-dotenv==1.0.0
langchain==0.1.0
openai==1.3.0
python-dateutil==2.8.2
